requirements.txt
aiohttp==3.8.5
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.4
flasgger==0.9.7b2
Flask==2.3.2
Flask-Admin==1.6.1
Flask-SQLAlchemy==3.0.5
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
minio==7.1.15
mistune==3.0.1
multidict==6.0.4
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.9
pydantic==2.0.2
pydantic-settings==2.0.3
pydantic_core==2.1.2
pypng==0.20220715.0
python-dotenv==1.0.0
pytz==2022.1
PyYAML==6.0.1
qrcode==7.4.2
referencing==0.30.0
requests==2.31.0
rpds-py==0.9.2
six==1.16.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
ujson==5.8.0
urllib3==2.0.4
uuid==1.30
Werkzeug==2.3.6
WTForms==3.0.1
yarl==1.9.2
YooMoney==0.1.0
